+++ packages/playground/src/types/utils.ts
@@ -1,0 +1,3 @@
+export type HasKey = {
+  [key: string]: any;
+};