+++ packages/vpdf/src/components/index.ts
@@ -1,12 +1,11 @@
 import { defineAsyncComponent } from "vue";
-import { usePdf, usePdfViewer } from "../composables";
 
-export { usePdf, usePdfViewer };
+export const VPdf = defineAsyncComponent(() => import("./vPdf.vue"));
+export const VPdfPage = defineAsyncComponent(
+  () => import("./vPdfPageRenderer.vue"),
+);
+export const VPdfViewer = defineAsyncComponent(
+  () => import("./vPdfViewer.vue"),
+);
 
-export default {
-  VPdf: defineAsyncComponent(() => import("./vPdf.vue")),
-  VPdfPageRenderer: defineAsyncComponent(
-    () => import("./vPdfPageRenderer.vue")
-  ),
-  VPdfViewer: defineAsyncComponent(() => import("./vPdfViewer.vue")),
-};
+export default { VPdf, VPdfPage, VPdfViewer };