+++ packages/vpdf/src/components/menu/pager.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { computed, defineAsyncComponent, ref, watch } from "vue";
+import { uniqid } from "../../composables";
 
 const Button = defineAsyncComponent(() => import("../button/index.vue"));
 
@@ -9,12 +10,15 @@ const props = withDefaults(
     pages: number;
     disabled?: boolean;
     mode: "vertical" | "horizontal";
+    label?: string;
+    hidePage?: boolean;
+    hideNavigation?: boolean;
   }>(),
   {
     modelValue: 1,
     pages: 0,
     disabled: false,
-  }
+  },
 );
 
 const emit = defineEmits(["update:modelValue"]);
@@ -22,6 +26,8 @@ const emit = defineEmits(["update:modelValue"]);
 const p = ref(props.modelValue);
 
 const inputSize = computed<number>(() => Math.max(`${props.pages}`.length, 3));
+
+const pagerId = uniqid();
 
 const isInt = (n: number) => n % 1 === 0;
 const onChange = (e: Event) => {
@@ -43,59 +49,60 @@ watch(
   () => props.modelValue,
   (val) => {
     p.value = val;
-  }
+  },
 );
 </script>
 <template>
-  <div class="flex items-center justify-center gap-2 px-2">
-    <div class="flex items-center gap-0.5">
+  <div
+    v-if="!hidePage || !hideNavigation"
+    class="vpdf:flex vpdf:items-center vpdf:justify-center vpdf:gap-2 vpdf:px-2"
+  >
+    <div
+      v-if="!hideNavigation"
+      class="vpdf:hidden vpdf:items-center vpdf:gap-0.5 vpdf:md:!flex"
+    >
       <Button
         @click="changePage(-1)"
         icon="arrow_circle_up"
-        class="group rounded-lg px-1 py-0.5 text-2xl"
-        :iconClass="{
-          'transition-transform': true,
-          '-rotate-90': mode != 'vertical',
-        }"
+        class="vpdf:transition-transform"
+        :iconClass="{ 'vpdf:-rotate-90': mode != 'vertical' }"
         :disabled="disabled"
       />
       <Button
         @click="changePage(1)"
         icon="arrow_circle_down"
-        class="group rounded-lg px-1 py-0.5 text-2xl"
-        :iconClass="{
-          'transition-transform': true,
-          '-rotate-90': mode != 'vertical',
-        }"
+        class="vpdf:transition-transform"
+        :iconClass="{ 'vpdf:-rotate-90': mode != 'vertical' }"
         :disabled="disabled"
       />
     </div>
     <div
-      class="hidden select-none items-center justify-stretch gap-1 md:flex"
+      v-if="!hidePage"
+      class="vpdf:flex vpdf:items-center vpdf:justify-stretch vpdf:gap-1 vpdf:select-none"
       :class="{
-        'text-gray-400': disabled,
+        'vpdf:text-background-400': disabled,
       }"
     >
       <label
-        for="__PDF__PAGER_INPUT__"
-        class="group relative max-w-10 bg-foreground/10 leading-none before:absolute before:inset-x-0 before:top-full before:h-0.5 before:bg-gray-400/25"
-        :class="{ 'cursor-not-allowed opacity-50': disabled }"
+        :for="pagerId"
+        class="vpdf:group vpdf:relative vpdf:max-w-10 vpdf:min-w-10 vpdf:bg-foreground/10 vpdf:leading-none vpdf:before:absolute vpdf:before:inset-x-0 vpdf:before:top-full vpdf:before:h-0.5 vpdf:before:bg-background-400/25"
+        :class="{ 'vpdf:cursor-not-allowed vpdf:opacity-50': disabled }"
       >
         <input
-          id="__PDF__PAGER_INPUT__"
+          :id="pagerId"
           :value="p"
           :size="inputSize"
-          class="w-full select-none bg-transparent px-1 text-center leading-none outline-none"
-          :class="{ 'aria-disabled pointer-events-none': disabled }"
+          class="vpdf:w-full vpdf:bg-transparent vpdf:px-1 vpdf:text-center vpdf:leading-none vpdf:outline-none vpdf:select-none"
+          :class="{ 'vpdf:aria-disabled vpdf:pointer-events-none': disabled }"
           :disabled="disabled"
           @change="onChange"
         />
         <div
-          class="absolute left-1/2 top-full h-0.5 w-0 -translate-x-1/2 bg-blue-600 opacity-0 transition-all group-focus-within:w-full group-focus-within:opacity-100"
+          class="vpdf:absolute vpdf:top-full vpdf:left-1/2 vpdf:h-0.5 vpdf:w-0 vpdf:-translate-x-1/2 vpdf:bg-primary-600 vpdf:opacity-0 vpdf:transition-all vpdf:group-focus-within:w-full vpdf:group-focus-within:opacity-100"
         />
       </label>
-      <span class="leading-none">/</span>
-      <span class="leading-none">{{ pages }}</span>
+      <span class="vpdf:leading-none">/</span>
+      <span class="vpdf:leading-none">{{ pages }}</span>
     </div>
   </div>
 </template>