+++ vite.config.ts
@@ -1,6 +1,7 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
-import tailwindcss from "tailwindcss";
+import tailwindcss from "@tailwindcss/vite";
+import type { PreRenderedAsset } from "rollup";
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -15,14 +16,16 @@ export default defineConfig({
     emptyOutDir: true,
     rollupOptions: {
       output: {
-        chunkFileNames: `[hash].js`,
+        // entryFileNames: `assets/[hash].js`,
+        chunkFileNames: `assets/[hash].js`,
+        assetFileNames: (assetInfo: PreRenderedAsset) => {
+          if (assetInfo.names?.[0].endsWith(".css")) {
+            return "assets/style.css";
+          }
+          return assetInfo.names?.[0];
+        },
       },
     },
   },
-  plugins: [vue()],
-  css: {
-    postcss: {
-      plugins: [tailwindcss],
-    },
-  },
+  plugins: [vue(), tailwindcss()],
 });