+++ packages/vpdf/HISTORY.md
@@ -1,3 +1,22 @@
+### 2.0.0-alpha / 08-28-2025
+
+    - Upgrade to Taiwind 4
+        - Add class prefix to avoid conflicts
+        - Add custom color shades (50 - 950) (background, foreground, primary, secondary, accent, positive, negative)
+        - Customize colors using css variables
+        - Add variables to customize icons
+    - Change layout
+        -Offset some menu items to the bottom
+    - Update scaling logic
+        - Add `scale` property for `VPdf`
+            - Sets the scale on initial load
+        - Add negative scale factors
+            - -1 => Fit Page
+            - -2 => Fit Width
+            - -3 => Fit Height
+        - Add more hidden scale factors
+    - Update Readme
+
 ### 1.0.2-alpha / 03-11-2025
 
     - Fixed build issue